Many changes to the implementation
* use CMake 3.9 for OpenMP linking
* Change all floats to double for armadillo usage
* Change Ellipsoid2D to a 3D implementation

+++ include/DynamicObstacleAvoidance/Agent.hpp
@@ -31,22 +31,22 @@ public:
 		return this->state.get_pose();
 	}
 
-	inline const Eigen::Vector3f get_position() const 
+	inline const Eigen::Vector3d get_position() const 
 	{ 
 		return this->state.get_position();
 	}
 
-	inline const Eigen::Quaternionf get_orientation() const 
+	inline const Eigen::Quaterniond get_orientation() const 
 	{ 
 		return this->state.get_orientation();
 	}
 
-	inline const Eigen::Vector3f get_linear_velocity() const
+	inline const Eigen::Vector3d get_linear_velocity() const
 	{
 		return this->state.get_linear_velocity();
 	}
 
-	inline const Eigen::Vector3f get_angular_velocity() const
+	inline const Eigen::Vector3d get_angular_velocity() const
 	{
 		return this->state.get_angular_velocity();
 	}
@@ -61,22 +61,22 @@ public:
 		this->state.set_pose(pose);
 	}
 
-	inline void set_position(const Eigen::Vector3f& position)
+	inline void set_position(const Eigen::Vector3d& position)
 	{
 		this->state.set_position(position);
 	}
 
-	inline void set_orientation(const Eigen::Quaternionf& orientation)
+	inline void set_orientation(const Eigen::Quaterniond& orientation)
 	{
 		this->state.set_orientation(orientation);
 	}
 
-	inline void set_linear_velocity(const Eigen::Vector3f& linear_velocity)
+	inline void set_linear_velocity(const Eigen::Vector3d& linear_velocity)
 	{
 		this->state.set_linear_velocity(linear_velocity);
 	}
 
-	inline void set_angular_velocity(const Eigen::Vector3f& angular_velocity)
+	inline void set_angular_velocity(const Eigen::Vector3d& angular_velocity)
 	{
 		this->state.set_angular_velocity(angular_velocity);
 	}
